+++ src/plugins/pluginTypes.ts
@@ -1,5 +1,7 @@
+import type { ZodType } from "zod/v4";
+
 export interface BasePluginType {
-  config: unknown;
+  configSchema: ZodType;
   customOverrideCriteria: Record<string, unknown>;
   customOverrideMatchParams: Record<string, unknown>;
   state: any;